test: isolate Dock notification routing
